Allow to define methods within action and controller blocks.
We often need to create helper functions that we reuse between different
actions, but we couldn't define them within action or controller blocks,
so we ended up defining them at the top level namespace. This is bad
because it is possible to create a clash by creating two methods with
the same name.